FEAT #5661 start status administration MVC

+++ rest/index.php
@@ -105,11 +105,12 @@ $app->get('/administration/users', \Core\Controllers\UserController::class . ':g
 $app->get('/administration/user/{userId}', \Core\Controllers\UserController::class . ':getUserForAdministration');
 
 //status
-$app->get('/status', \Core\Controllers\StatusController::class . ':getList');
-$app->get('/status/{id}', \Core\Controllers\StatusController::class . ':getById');
-$app->post('/status', \Core\Controllers\StatusController::class . ':create');
-$app->put('/status', \Core\Controllers\StatusController::class . ':update');
-$app->delete('/status/{id}', \Core\Controllers\StatusController::class . ':delete');
+$app->get('/administration/status', \Core\Controllers\StatusController::class . ':getList');
+$app->get('/administration/status/lang', \Core\Controllers\StatusController::class . ':getLang');
+$app->get('/administration/status/{id}', \Core\Controllers\StatusController::class . ':getById');
+$app->post('/administration/status', \Core\Controllers\StatusController::class . ':create');
+$app->put('/administration/status', \Core\Controllers\StatusController::class . ':update');
+$app->delete('/administration/status/{id}', \Core\Controllers\StatusController::class . ':delete');
 
 //docserver
 $app->get('/docserver', \Core\Controllers\DocserverController::class . ':getList');
